allow manually choosing an engage file
This generally shouldn't be used since it would complicate invocation.
I can imagine some cases where it might be useful, though.

One thing you can do with this that's is hilarious but not recommended
is to add `#!/usr/bin/env engage -f` to the top of your Engage file and
make it executable. While it does totally work, it's... kinda weird.