disallow certain things as group names
This is for forward-compatibility. Ideally I'd like to remove the `just`
subcommand in favor of simply typing `engage <group> [task]`. With this
syntax however, it becomes impossible to invoke a group named `self` or
`help` since we already have a subcommand for that. If we leave `just`
in for backward compatibility, we'll have the same problem with that
too.