greatly improve error messages for dep cycles
Instead of a vague "there are errors lol" message, it now actually tells
you which nodes have edges that participate in the cycle, letting you
decide which edge(s) to break in order to make your graph acyclic.

Also, acyclic-ness is no longer required for the `self dot` subcommand,
so this can be used to help visually debug your graph.