rework the output while running the graph
The previous output format would lose the information of the original
stream the lines were coming from if colors were not visible. What's
worse is that the colors used were red and green, which is the most
common kind of colorblindness.

Now, the output looks a little prettier and does not rely on color to
encode information at all, but is purely decorative. In the output, each
line will contain an `O` or an `E` to indicate that it came from stdout
or stderr respectively.