get rid of the `self` subcommand
All of the subcommands that lived under `self` are now available at the
top level instead.

I originally had a `self` subcommand because I wanted to get rid of the
`just` subcommand and allow invocations like `engage group task` or
`engage group`, but it turns out this is difficult/impossible to parse
correctly and `clap` doesn't even attempt to support this.