give up on trying to synchronize processes exiting
Now we just replace the text that's nondeterministic. This is fine
because this test is mainly about asserting the formatting of the box
drawing characters and not about the error messages themselves, as we
have other tests for that already. This also happens to be a lot more
portable since it doesn't rely on OS-specific things to observe process
state.

I'm making this change because, to my surprise, this immediately
deadlocked in CI and also on @curid:matrix.org's machine.