rewrite `graph::execute` again
This fixes a bug where given a graph of nodes A, B, and C with an edge
from A to B, where A succeeds immediately and C sleeps forever, B is
never run despite its dependencies being satisfied. A test for this case
is now included.

The bug was introduced in cb470cebb8b2a2e06dad3c2ede8e453a25bfd169
and the fix is to more or less go back to how it was before. This
implementation is still an improvement though, particularly in how new
nodes to visit are calculated.